database_managerクラス作成（db処理を切り離し）
